+++ NEWS.md
@@ -1,6 +1,8 @@
 # qualpalr 0.4.2.9000
 
 ## Minor changes
+* `qualpal()` gains an argument, `n_threads`, for specifying the number
+of threads to use when computing the distance matrix between colors.
 * C++ functions call namespaces explicitly using `::`.
 
 # qualpalr 0.4.2